Added ProviderTypingTrainer and TestsTypingTrainerGoodCases. Wrote some tests to test the tests (lol) and for feedback.class.

+++ components/src/main/java/de/hhn/it/pp/components/typingtrainer/Feedback.java
@@ -32,7 +32,7 @@ public class Feedback {
     return time;
   }
 
-  public void setTime(float time) {
+  public void setTime(double time) {
     this.time = time;
   }
